Moved Renderer to renderers.py
Also renamed it HtmlRenderer. This makes code easier to understand and gets dir
ready for a new renderer that outputs csv
