+++ scripts/feature_toggle_report_generator.py
@@ -2,55 +2,14 @@
 
 import datetime
 import io
-import logging
 import os
 
 import click
 import jinja2
 
 from scripts.ida import IDA, add_toggle_state_to_idas, add_toggle_annotations_to_idas
+from scripts.renderers import HtmlRenderer
 
-LOGGER = logging.getLogger(__name__)
-logging.basicConfig(level=logging.INFO)
-
-
-class Renderer(object):
-
-    def __init__(self, template_dir, report_dir):
-        self.jinja_environment = jinja2.Environment(
-            autoescape=False,
-            loader=jinja2.FileSystemLoader('templates'),
-            lstrip_blocks=True,
-            trim_blocks=True
-        )
-        self.report_dir = report_dir
-        if not os.path.isdir(report_dir):
-            os.mkdir(report_dir)
-
-    def render_file(self, output_file_name, template_name, variables={}):
-        file_path = os.path.join(self.report_dir, output_file_name)
-        template = self.jinja_environment.get_template(template_name)
-        with io.open(file_path, 'w') as output:
-            output.write(
-                template.render(**variables)
-            )
-
-    def render_html_report(self, idas, environment_name, show_state=True):
-        report_date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M")
-        report_path = 'feature_toggle_report.html'
-        LOGGER.info('Attempting to render HTML report')
-        self.render_file(
-            report_path, 'report.tpl',
-            variables={
-                'idas': idas, 'environment': environment_name,
-                'report_date': report_date,
-                'show_state': show_state,
-
-            }
-        )
-        LOGGER.info(
-            'Succesfully rendered HTML report to {}'.format(report_path)
-        )
 
 
 @click.command()
@@ -78,7 +37,7 @@ def main(sql_dump_path, annotation_report_path, output_path, environment_name, s
     if show_state:
         add_toggle_state_to_idas(idas, sql_dump_path)
     add_toggle_annotations_to_idas(idas, annotation_report_path)
-    renderer = Renderer('templates', output_path)
+    renderer = HtmlRenderer('templates', output_path)
     renderer.render_html_report(idas, environment_name, show_state)
     if publish:
         confluence = create_confluence_connection()
